you can now ignore commands from channels or members

+++ bot.py
@@ -143,19 +143,16 @@ class Moderation(commands.Cog):
     
     async def bot_check(self, ctx):
         """Checks if the message calls a command. If the channel is disabled or the member is banned from using commands, this will raise an error"""
-        if ctx.command is not None:
-            async with aiosqlite.connect("databases/main.db") as db:
-                cursor = await db.execute("SELECT * FROM ignored_channels WHERE channel_id = ?",(ctx.channel.id,))
-                res = await cursor.fetchone()
-            if res:
-                is_channel_ignored = True
-            else:
-                is_channel_ignored = False
-            if is_channel_ignored:
-                raise commands.DisabledCommand("You can't use commands in this channel.")
-            return True
-        elif not ctx.guild:
-            raise commands.NoPrivateMessage("I can't be used in DM. Please, use this command in another channel.")
+        async with aiosqlite.connect("databases/main.db") as db:
+            cursor = await db.execute("SELECT * FROM ignored_channels WHERE channel_id = ?",(ctx.channel.id,))
+            res_channel = await cursor.fetchone()
+            cursor = await db.execute("SELECT * FROM ignored_members WHERE member_id = ? AND guild_id = ?",(ctx.author.id,ctx.guild.id))
+            res_member = await cursor.fetchone()
+        if res_channel:
+            raise commands.DisabledCommand("You can't use commands in this channel.")
+        if res_member:
+            raise commands.DisabledCommand("You are not allowed to use commands on this server.")
+        return True
 
     @commands.Cog.listener()
     async def on_guild_join(self,guild:discord.Guild):
@@ -256,26 +253,26 @@ class Moderation(commands.Cog):
             embedVar.add_field(name=f"You were banned from {ctx.guild} by {ctx.author}.",value=f"Reason : {reason}")
             embedVar.set_footer(text=f"Requested by {ctx.author}.")
             await member.send(embed=embedVar)
-            await member.ban(reason=reason)
+            await member.ban(reason=reason) 
 
     @ban.command()
-    async def match(ctx,r,*,words):
+    async def match(self,ctx,reason,*,words):
         """Bans every member that said [words] in the last 100 messages of this channel. Reason must be quoted if it isn't a single word"""
-        banned_words = words.split(" ")
-        last_100_messages = await ctx.channel.history(limit=123).flatten()
+        banned_words = words.split(",") #Split the words if there are more than one
+        last_100_messages = await ctx.channel.history(limit=123).flatten() #Is a list of last 100 messages from this channel.
         count = 0
         for message in last_100_messages:
-            if message.author == ctx.author or message.author.bot:
+            if message.author == ctx.author or message.author.bot: #Bots won't be banned, and the person who called this command won't either. So go to the next message
                 continue
             for word in banned_words:
-                if word in message.content:
+                if word in message.content: #If a banned word is inside the message
                     embedVar = discord.Embed(title="Uh oh. Looks like you did something QUITE bad !",color=0xff0000)
-                    embedVar.add_field(name=f"You were banned from {ctx.guild} by {ctx.author}.",value=f"Reason : {r}")
+                    embedVar.add_field(name=f"You were banned from {ctx.guild} by {ctx.author}.",value=f"Reason : {reason}")
                     embedVar.set_footer(text=f"Requested by {ctx.author}.")
                     await message.author.send(embed=embedVar)
                     await message.author.ban(reason=reason)
                     count += 1
-                    break
+                    break #No needs to know if there are two words that could make you banned in the message. One is enough
         if count:
             await ctx.send((f'Done ! I banned {count} people.' if count > 1 else "Done ! I banned one person."))
         else:
@@ -283,6 +280,7 @@ class Moderation(commands.Cog):
 
     @ban.group(invoke_without_command=True)
     async def time(self,ctx,member:discord.Member,time:TimeConverter,*,reason="Not specified."):
+        """Basically a ban command where you can add a duration. Once it's over, the bot automatically unbans the member"""
         if ctx.invoked_subcommand is None:
             embedVar = discord.Embed(title="Uh oh. Looks like you did something QUITE bad !",color=0xff0000)
             embedVar.add_field(name=f"You were banned from {ctx.guild} by {ctx.author}.",value=f"Reason : {reason}")
@@ -294,10 +292,10 @@ class Moderation(commands.Cog):
     
     @time.command()
     async def match(ctx,time:TimeConverter,reason,*,words):
-        """Bans every member that said [words] in the last 100 messages of this channel. Reason must be quoted if it isn't a single word"""
+        """Bans every member that said [words] in the last 100 messages of this channel for a [time] duration. Reason must be quoted if it isn't a single word"""
         banned_words = words.split(" ")
         last_100_messages = await ctx.channel.history(limit=123).flatten()
-        banned_people = []
+        banned_people = [] #will store discord.Member objects
         for message in last_100_messages:
             if message.author == ctx.author or message.author.bot:
                 continue
@@ -316,15 +314,16 @@ class Moderation(commands.Cog):
         else:
             await ctx.send("No one said those awful words.")
         await asyncio.sleep(time)
-        for i in banned_people:
+        for i in banned_people: #Once the ban duration is over, iterate through the list of banned people and then unban them.
             await i.unban()
 
 
     @commands.command(aliases=["u","unbna"])
     @commands.has_permissions(ban_members = True)
     async def unban(self,ctx,person,*,reason="Not specified."):
+        """Unbans the member from the guild"""
         bans = await ctx.guild.bans()
-        if len(bans) == 0:
+        if len(bans) == 0: #Checks if the banlist is empty.
             embedVar = discord.Embed(title="Uh oh. Looks like no one is banned on this server. Those are good news !",color=0xaaffaa)
             return await ctx.send(embed=embedVar)
         elif person == "all":
@@ -382,7 +381,7 @@ class Moderation(commands.Cog):
     @commands.has_permissions(administrator = True)
     async def perms(self,ctx,member:discord.Member):
         embedVar = discord.Embed(title=f"You asked for {member}'s permissions on {ctx.guild}.",color=0xaaaaff)
-        embedVar.add_field(name="Here they are : ",value="\n".join(["• {}".format(i[0]) for i in member.guild_permissions if i[1] is True]))
+        embedVar.add_field(name="Here they are : ",value="\n".join(["• {}".format(i[0]) for i in member.guild_permissions if i[1] is True])) #Iterate through the discord.Member permissions on the guild. If they have the permission, this is added to the list.
         await ctx.author.send(embed=embedVar)
 
     @commands.group(invoke_without_command=True)
@@ -431,13 +430,33 @@ class Moderation(commands.Cog):
         await member.unban(reason="Softban.")
         await ctx.send(f"{member} was softbanned.")
 
-    @commands.group()
+    @commands.group(invoke_without_command=True)
     async def ignore(self,ctx):
         if ctx.invoked_subcommand is None:
             await ctx.send("Use the 'member' or 'channel' subcommand to specify who shouldn't be allowed to use a command, or where people won't be allowed to use those beautiful commands !")
-    
+
+    @ignore.command(name="member")
+    async def _m(self,ctx,member:discord.Member=None):
+        async with aiosqlite.connect("databases/main.db") as db:
+            cursor = await db.execute("SELECT member_id FROM ignored_members WHERE member_id = ? AND guild_id = ?",(member.id,ctx.guild.id))
+            result = await cursor.fetchone()
+            if result:
+                await ctx.send("This member is already ignored. Type 'yes' if you want them to be able to use commands on this server !")
+                try:
+                    answer = await self.bot.wait_for("message",check=lambda m: m.author == ctx.author and m.channel == ctx.channel,timeout=10)
+                except asyncio.TimeoutError:
+                    await ctx.send("You didn't answer fast enough. Aborting the process !")
+                else:
+                    await db.execute("DELETE FROM ignored_members WHERE member_id = ? AND guild_id = ?",(member.id,ctx.guild.id))
+                    await db.commit()
+                    await ctx.send(f"{member.mention} now has enabled commands !")
+            else:
+                await db.execute("INSERT INTO ignored_members VALUES(?,?);",(member.id,ctx.guild.id))
+                await db.commit()
+                await ctx.send(f"{member.mention} now has disabled commands !")
+
     @ignore.command()
-    async def channel(self,ctx,channel:discord.TextChannel):
+    async def channel(self,ctx,channel:discord.TextChannel=None):
         async with aiosqlite.connect("databases/main.db") as db:
             cursor = await db.execute("SELECT channel_id FROM ignored_channels WHERE channel_id = ?",(channel.id,))
             result = await cursor.fetchone()
@@ -448,16 +467,18 @@ class Moderation(commands.Cog):
                 except asyncio.TimeoutError:
                     await ctx.send("You didn't answer fast enough. Aborting the process !")
                 else:
-                    await db.execute("DELETE FROM ignored_channels WHERE channel_id = ?",(channel.id,))
-                    await db.commit()
-                    await ctx.send(f"{channel.mention} now has enabled commands !")
+                    if answer.content.lower() == "yes":
+                        await db.execute("DELETE FROM ignored_channels WHERE channel_id = ?",(channel.id,))
+                        await db.commit()
+                        await ctx.send(f"{channel.mention} now has enabled commands !")
+                    else:
+                        await ctx.send("Aborting process !")
             else:
                 await db.execute("INSERT INTO ignored_channels VALUES(?);",(channel.id,))
                 await db.commit()
                 await ctx.send(f"{channel.mention} now has disabled commands !")
 
-
-    @commands.group(invoke_without_command=True,name="purge")
+    @commands.group(invoke_without_command=True)
     async def purge(self,ctx,Amount:int=2): #Delete "Amount" messages from the current channel. $purge [int]
         await ctx.message.delete()
         if ctx.invoked_subcommand is None:
@@ -626,8 +647,6 @@ class Tags(commands.Cog):
         
     @tag.command()
     async def add(self,ctx,tag_name,*,description):
-        print(f"Tag name : {tag_name}")
-        print(f"Description : {description}")
         guild_id = f"_{ctx.guild.id}"
         async with aiosqlite.connect("databases/tags.db") as db:
             async with db.execute(f"SELECT description FROM {guild_id} WHERE tag_name = ?;",(tag_name,)) as cursor:
